mini-porjects/03-medallion-architecture/retailapp/requirements.txt
psycopg2==2.9.9
python-dotenv==1.0.1
setuptools==69.5.1
wheel==0.43.0
